Conflicts:
	MandelbrotSetVis.xcodeproj/project.xcworkspace/xcuserdata/esie.xcuserdatad/UserInterfaceState.xcuserstate
	MandelbrotSetVis/Mandelbrot/RendererBuffer.swift
	MandelbrotSetVis/MetalRenderer/Bridging/SwiftToMetalConvertible.swift

+++ MandelbrotSetVis/Mandelbrot/RendererBuffer.swift
@@ -9,5 +9,5 @@ struct RendererBuffer {
     var scale: Float32 = 1.0
     var translation: (x: Float32, y: Float32) = (0.0, 0.0)
     var aspectRatio: (x: Float32, y: Float32) = (1.0, 1.0)
-    var iterations: Float32 = 256.0
+    var iterations: Int = 256
 }
